exercises(raindrops): use u32, not usize
Make it more reasonable to add a test for a large number later, and make
the required buffer size more obvious (the maximum value of a usize on a
64-bit platform has 20 digits).

+++ exercises/practice/raindrops/.meta/example.zig
@@ -1,6 +1,6 @@
 const std = @import("std");
 
-pub fn convert(buffer: []u8, comptime n: usize) []const u8 {
+pub fn convert(buffer: []u8, comptime n: u32) []const u8 {
     const pling = if (n % 3 == 0) "Pling" else "";
     const plang = if (n % 5 == 0) "Plang" else "";
     const plong = if (n % 7 == 0) "Plong" else "";
