+++ exercises/practice/raindrops/raindrops.zig
@@ -1,4 +1,4 @@
-pub fn convert(buffer: []u8, n: usize) []const u8 {
+pub fn convert(buffer: []u8, n: u32) []const u8 {
     _ = buffer;
     _ = n;
     @compileError("please implement the convert function");
